requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic_core
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
SQLAlchemy
pysqlite3
pymysql
alembic
async-cache
#python-magic-bin
asyncer